requirements.txt
akshare>=1.13.25
pandas>=2.1.0
numpy>=1.26.0
yfinance>=0.2.40
requests>=2.31.0
PyYAML>=6.0.1
pdfplumber>=0.10.3
pdfminer.six>=20221105
beautifulsoup4>=4.12.2
lxml>=4.9.3
